[Android/Feat] 루틴 등록 기능 구현 - 조명 S12P31D103-185

See merge request s12-final/S12P31D103!136

+++ Android/app/src/main/java/com/hogumiwarts/lumos/ui/screens/control/light/LightIntent.kt
@@ -2,9 +2,9 @@ package com.hogumiwarts.lumos.ui.screens.control.light
 
 
 sealed class LightIntent {
-    data class LoadLightStatus(val deviceId: Long): LightIntent()
+    data class LoadLightStatus(val deviceId: Int): LightIntent()
 
-    data class ChangeLightPower(val deviceId: Long, val activated: Boolean): LightIntent()
+    data class ChangeLightPower(val deviceId: Int, val activated: Boolean): LightIntent()
 
     data class ChangeLightBright(val deviceId: Long, val brightness: Int): LightIntent()
